feat - move class, Bank
