+++ src/main/java/co/invest72/exchange_rate/infrastructure/api/KoreaeximProperties.java
@@ -1,5 +1,7 @@
 package co.invest72.exchange_rate.infrastructure.api;
 
+import java.util.Objects;
+
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 @ConfigurationProperties(prefix = "external-api.koreaexim")
@@ -8,7 +10,7 @@ public class KoreaeximProperties {
 	private final String baseUrl;
 
 	public KoreaeximProperties(String apiKey, String baseUrl) {
-		this.apiKey = apiKey;
-		this.baseUrl = baseUrl;
+		this.apiKey = Objects.requireNonNull(apiKey);
+		this.baseUrl = Objects.requireNonNull(baseUrl);
 	}
 }
