+++ src/main/java/co/invest72/financial_product/application/FinancialProductService.java
@@ -13,7 +13,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import co.invest72.common.time.LocalDateProvider;
-import co.invest72.exchange_rate.domain.Bank;
+import co.invest72.exchange_rate.domain.service.Bank;
 import co.invest72.financial_product.domain.FinancialProduct;
 import co.invest72.financial_product.domain.FinancialProductRepository;
 import co.invest72.financial_product.domain.entity.FinancialProductData;
